Add rule runner
Added a new helper that should make testing rules a lot easier. For now it only handles HTML fixtures (`element::nodename`) but expanding should be fairly easy.
Now a developer only needs to specify the name of the test, the events to receive (and where to get the data) and what is the expected output when testing a rule.

* Refactor `lang-attribute` rule to new test runner
* Create rule runner and types (some could use a better name)