Refactor API to be async
Chrome Debugging Protocol is an async protocol so we need to update our API
for all collectors to be async as well. This is the first cut to it. It is
expected to be more changed once the development of CDP continues.
Some of the changes:
* Resources now need their own folder (this wasn't the case for collectors and formatters)
The reason is that collectors might require extra helper files and this makes the
organization cleaner.
* `AsyncHTMLElement` and `AsyncHTMLDocument`: they expose the minimum pieces for the rules
and helpers to run taking into account what methods are `async` on CDP. This way we should
prevent also unwanted modifications to the DOM.
* Update all code to work with the new code.