Refactor rule-runner
rule-runner now creates a fully working `sonar` object with a given `collector` (for now
only `jsdom` but the idea is to automatically tests with all the available ones) and just
the rule to be tested. To achieve this, a new `test-server` has been added. This way we can
write even more concise tests. Check the section `How to test rules` in the `readme.md` for
more information about it and how to use it.

The current tests has been updated to use this new method.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Fix #61