Close CDP browser when process finishes
Close all the tabs that are being used by the `Sonar` object.
If no other tabs are open, the browser will close.
If there are remaining tabs not created by the `Sonar`
object, then it will not force the exit. Other `Sonar`s or the
user might be using the browser.

* Rename `collectors.ts` to `collector.ts`.
* Add `close` to `collector.ts`.
* Store the `tab` and close it on `close`.
* Close the CDP `client` on `close`.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Fix #103