Close tabs and concurrency improvements with CDP
Tests with `ava` are run in parallel so relying only in `cdp.pid`
is not possible because several tests might try to open chrome at
the same time and cause some issues. To solve this we need to use
[`lockfile`](https://github.com/npm/lockfile) with some aggressive
number for retry, poll, etc. These numbers might need to be
tweaked later if `travis` causes some trouble.

* Use `lock` to open a new browser instance and wait if needed.
* Correctly open a new tab if needed.
* Close tabs `async`ly on `close`.
* Call to `close` where needed.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Partial fix of #113