Change `location` of issues
Per [this comment](https://github.com/MicrosoftEdge/Sonar/issues/63#issuecomment-295436952)
we are changing to just report the issue relative to the element
and let `rule-context` (in the future) figure out the rest.
This unblocks us to continue testing `collector`s and make sure
they are consistent at least in that aspect.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Work on #63
Work on #113