Fix: Issue with Cloudinary and error handling
The cloudinary npm package doesn't seem to like streams and invalid
content. Even if the error is handled in the callback, there is an
internal `Promise` that is never returned and is `reject`ed and thus
making the whole process fail.

This fix replaces streams with real files, which seems to be the
better handled scenario.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Ref: [Promise not always returned][promise-return]
Ref: [Upload using promises][promise-upload]
Ref: [Upload using Buffer][buffer-upload]

[buffer-upload]: https://github.com/cloudinary/cloudinary_npm/issues/130#issuecomment-308444514
[promise-return]: https://github.com/cloudinary/cloudinary_npm/blob/0a54a2db5c50b81c4f72dde8b97fccf66ecfcb7b/src/uploader.coffee#L271
[promise-upload]: https://github.com/cloudinary/cloudinary_npm/issues/139#issuecomment-290024896