Fix: Downgrade HTTP/2 headers in `rawResponse`
`rawResponse` in `connector`s that use the debugging protocol rely
on `requester`. Unfortunatelly the `request` package doesn't support
HTTP/2 and we were passing the original request headers that could
contain `Pseudo-Header Fields` and thus causing an exception.

The solution is to downgrade the headers to valid HTTP/1.1 by removing
all the `Pseudo-Header Fields`.