Chore: Ignore injected code in coverage to fix `yarn test`
When running code coverage in `yarn test` code in the file is
modified to include statements like `cov_l3mx4wetg.s[46]++;`
which fail in `injectedCode` since it runs in a different context.

Adding `/* istanbul ignore next */` before `injectedCode`
disables code coverage for this function, fixing the issue.
Although that means we don't get code coverage here, it is no
different than using a string of embedded script which is the
alternative used in other rules.

Documentation for disabling coverage:
https://www.npmjs.com/package/nyc#parsing-hints-ignoring-lines