Fix: Drop `mime-db` dependency
`mime-db` was used as a source to check if something should be
compressed, but the project is planning to drop the `compressible`
data in the near future in favor of the `compressible`¹ project.

All of the cases currently covered by `compressible` are already
covered by our code, so this change just adds the cases that were
covered by `mime-db` and not by our code, hence, maintaining
backwards compatibility.

Also, since `mime-db` is no longer used, this change brings, in some
cases, a minor performance improvement.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

¹ https://github.com/jshttp/compressible

Fix #1108
Close #1126