Chore: Enable auto-merge greenkeeper
The script `greenkeeper-merge.js` is executed as a job when all the
tests passed. `greenkeeper` does a commit per each package in the
monorepo that shares the same dependency. This scripts:

* Squashes all of those commits (takes all the commits from the end
  with the same message until one is different).
* Rebases the changes into master and pushes.

**NOTE**: The `install` step of `travis.yml` performs a `yarn`. If the
version is different, the generated `yarn.lock` file will be too. This
script discards any changes in the `yarn.lock` that originate in that
step (`greenkeeper` takes care of updating it).

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Fix #938