Fix: Disable hints that are currently noisy when run locally
Disabling `hint-apple-touch-icons` because it is not aware of if
the project is a PWA or not (needs to check for a manifest).

Disabling `webpack-config/module-esnext-typescript` because it
requires loading both the webpack and typescript configs at the
same time to work (which doesn't currently happen in vscode).

These can be re-enabled once the affected hints are updated.