Fix: Avoid exception if resource is not compressed
Check if the resource has the `content-enconding` header with a
value that can be processed. If not, use the resource's size
already available instead of trying to get the `rawResponse`.

This avoids going through the `exception` path in resources such
as images and thus improving the speed of the rule.