Build: Use TypeScript project references
Reduce build time for the project by ~50% and automatically
build dependencies for sub-packages. E.g. building
`connector-local` will build both `hint` and `parser-html` if needed.

The build order of sub-packages no longer needs to be specified in
the root `package.json` as `tsc -b` will use the provided references
to automatically determine the build order instead.

A new script, `update-tsconfig-references.ts` is included to create
references in each package's `tsconfig.json` file based on
dependencies in the associated `package.json`. This can be built
and run from the project root using `yarn update:references`, but
only needs to be run when new packages are added or cross-
package dependencies change.