Fix: Make `canvas` optional
`canvas` is not a must-have dependency. Users might use another
connector other than `jsdom` and even if they use this one, they can
still get results (images will not be downloaded thought). This should
fix a few issues people run into when installing for the first time the
project, especially if the binaries are not available.