Chore: Linting errors with latest `eslint-typescript-parser`
The work in db92e295f07c74c9e9315259e521929c6de6709a was not enough
and is still causing some issues. This PR refactors all destructuring
that confused the parser to be in the parameters, avoiding issues.
The rule `no-shadow` had to be disabled or otherwise ESLint reported
other errors as well.