Fix: Prevent loading `canvas` from `html-parser`
The `canvas` package is a source of problems, especially when there is a
version mismatch between the binary and the node version being executed.
This problem happens quite often with the VS Code extension where the
node version used could be behind from the latest one by several
releases.

In order to prevent that, the `html-parser` "hijacks" `require`'s cache
by placing a "fake" `canvas` module that does not crash `jsdom`
execution and allows the code execution to continue normally.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Fix #1482