Docs: Fix reference links and remove `markdownlint-cli` dependency
Create a custom rule for `markdownlint` that validates that a link
reference is valid and replace the calls to `markdownlint-cli` to the
new script in all packages.

`markdownlint-cli` does not have a way to add custom rules, you need to
use the API to do so and thus `markdownlint-cli` is no longer needed as
a dependency.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Ref: #1436