Fix: Gracefully handle invalid browserslist and URL queries
Fix #1666

Prevent form submission when analysis is run with an invalid
`browserslist` query. Similarly, prevent form submission
when an invalid `RegExp` is provided for ignored URLs.

Update the analysis script gracefully handle invalid should
any be successfully submitted by falling back to defaults.

Also provide feedback to the user during query construction on
whether or not the query is valid and why.