Chore: Remove obsolete failing test
The `keyframes` property is supported by Opera 15, but was
previously incorrectly reported as unsupported due to a bug in
the MDN data. This bug was fixed in
https://github.com/mdn/browser-compat-data/pull/3227.

No other un-prefixed CSS feature exists which is both included
in the MDN data and has been removed by a browser so removing
this test.