Fix: Fully exit Chrome when all tabs are closed
On macOS Chrome does not fully exit even if all the tabs are closed.
This causes a timeout when running the tests with `nyc` of about 2
minutes (or if the browser is manually closed). This addresses this
issue.
Because during testing multiple workers are using the same instance of
Chrome, one could try to close the browser while another is opening a
new tab and thus causing communication issues. The solution is to use
the same `lock` to launch the browser and to close it with a twist: The
`close` method gets all the possible Targets of type `page` (tabs),
close the ones related to the current session and see if there are any
other left. If there are not, that means it is safe to completely close
the browser by sending a `SIGKILL` (no other methods works because the
debugging protocol `client` cannot communicate in this state and
gracefully close the browser).
This issue was easier to reproduce on `hint-compat-api` because of the
ammount of separate tests being run.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Fix #2197