New: Allow to set default timeout to `page`
The `puppeteer` option `timeout` is only used to launch the browser. If
one wants to increase the timeout for a navigation or any action in
general, it has to be done via `page.setDefaultTimeout` or
`page.setDefaultNavigationTimeout`. This changes takes the value of
`puppeteer.options.timeout` and calls `page.setDefaultTimeout` with it.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Ref https://pptr.dev/#?product=Puppeteer&version=v1.16.0&show=api-pagesetdefaulttimeouttimeout
Ref https://pptr.dev/#?product=Puppeteer&version=v1.16.0&show=api-puppeteerlaunchoptions