Build: Enable incremental builds
Store `tsconfig.tsbuildinfo` in a different directoy than `dist` to
avoid removing it when running `npm run clean`.

With this change running `yarn build` shows the following numbers:

* Cold build (no `.tsbuildinfo`): 786s
* Incremental build: 366s

Raw TypeScript time should be even better as `yarn build` does more
than just compilinng.

It is not necessary to add `"incremental": true` to the base
`tsconfig.json` because it is enabled by default with `composite`
configurations, but it is added here for clarity of intention.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Fix #2078