Fix: Correctly find resources when runnning globally
When running globally (`npx` or `npm create`) the resources where
searched in the global installation folder instead of relative to where
the process is being run. This caused tools such as `npm create hintrc`
to never find any of the resources preinstalled which is not the
expected behavior.

While this might be considered a breaking change, global run with hints
outside the `hint` global folder never worked correctly.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

Fix #2263