Fix: Test for extension-vscode
The tests for extension vscode can fail because after the first time you
run yarn test, the map info is removed from the JavaScript files.
Then if you try to compile the files again, if anything has changed,
the compiler will not change the file and the map info wouldn't be
regenerated.

To solve this, we need to clean the old files and compile again.