New: Introduce new disallowed HTTP response headers
This change adds the `Expires`, `Host`, `P3P`, `Pragma`, `Via` and
`X-Frame-Options` headers to the `hint-no-disallowed-headers` hint.
This was based on advices from
https://www.fastly.com/blog/headers-we-dont-want.

These hints are introduced with specific warning messages, rather than
by being added to the general list of headers already present in the
hint.

Specifically, this change moves some of the logic from the constructor
to other places to make it smaller and easier to read.
It renames the `disallowedHeaders` property to something more
explanatory and introduces a new special list of disallowed headers,
each with their own warning message.