Fix: Better wrapping logic for config page
Using CSS grid minmax, we can make the various columns
in the config page be at least some predefined width that
avoids poor wrapping of form fields, and at most take all
the size they need.

Now, to avoid them stretching too wide and becoming hard to
read, I preserved the max-width that was already there, but
moved it to the parent element to avoid conflicting with
the auto-fit logic (see below), as this was causing overflow.

We're using auto-fit to make the number of columns adapt to
the available space.