New: Internet Explorer Flexbox Compat hint
Hint that warns users when they do configure IE as a target browser and
flexbox is used in one of the stylesheets.

This is the initial commit, with the boilerplate code only.
More work is needed on the message, and on adding tests.
This is just to get the ball rolling.

Fix: #4072