Fix: Restore handling relative file paths in getAsUri
This was regressed when attempting to simplify dependencies.
An incorrect assumption was made about passed paths already
being absolute, but in the CLI relative paths can make it here.

Fixed by resolving the passed paths before converting to a URL.
Also added a new test to validate relative paths are converted.

- - - - - - - - - - - - - - - - - - - -

Fix #4836