Chore: Fix test coverage issue by ensuring one test is headless
In CI all tests are headless, but locally most run showing the
browser UI. This caused getFavIcon to not be called as it's only
needed when running in headless mode. This in turn caused branch
coverage to be just short of our 80% goal when running locally.

Fixed by ensuring at least one test runs in headless mode even
locally so this branch of code is executed.