Fix: Avoid changing original snyk-snapshot.json on build
Changes the step which prunes the data included in the snapshot
to output directly to the dist folder instead of src folder.
Also updates the copy step to avoid copying the original
snapshot to dist.

This fixes the pruning step from updating the tracked original
file which was resulting in diffs in git (and occasionally made
it into PRs).