docs: clarify case-sensitive value check for highest-available-document-mode
Fixes #3506

Added a clear note in the "What does the hint check?" section to inform users that the value check is case-sensitive (ie=edge vs IE=edge). This addresses the confusion mentioned in issue #3506 where the case-sensitivity requirement was buried in the configuration section and not immediately obvious to users.